+++ src/scan_date_time.hh
@@ -58,7 +58,7 @@ struct ScanDateTime
     }
 
     std::string scan_date_time;
-    int scan_seconds;
+    long long scan_seconds;
 };
 
 std::string to_string(const ScanDateTime& scan_date_time);
